+++ cypress/locators/decimal/locators.js
@@ -2,5 +2,4 @@
 export const INPUT_PRECISION_SLIDER = 'input[name="precision"][type="range"]';
 
 // component preview locators
-export const LABEL_INPUT_PREVIEW = '[data-element="label"]';
 export const DECIMAL_PREVIEW = '[data-component="decimal"]';