+++ src/__experimental__/components/date-range/date-range.spec.js
@@ -107,7 +107,7 @@ describe('DateRange', () => {
         expect(customOnBlur).not.toHaveBeenCalled();
       });
 
-      it('does not fire an onBlur event when the endtDate is focused', () => {
+      it('does not fire an onBlur event when the endDate is focused', () => {
         wrapper.instance().focusEnd();
         wrapper.instance()._onBlur();
         expect(customOnBlur).not.toHaveBeenCalled();
@@ -368,6 +368,33 @@ describe('DateRange', () => {
       endInput = wrapper.find(BaseDateInput).at(1);
       expect(startInput.props().value).toEqual('2016-10-10');
       expect(endInput.props().value).toEqual('2016-11-11');
+    });
+
+    it('supports value update dynamically at runtime', () => {
+      wrapper = renderDateRange({
+        onChange: customOnChange,
+        value: ['2012-12-12', '2012-12-13']
+      }, mount);
+      wrapper.setProps({ value: ['2016-10-10', '2016-11-11'] });
+      wrapper.update();
+      startInput = wrapper.find(BaseDateInput).at(0);
+      endInput = wrapper.find(BaseDateInput).at(1);
+      expect(startInput.props().value).toEqual('2016-10-10');
+      expect(endInput.props().value).toEqual('2016-11-11');
+    });
+
+    it('supports value update dynamically at runtime and defaults to today if new value is undefined', () => {
+      MockDate.set('2020-01-21');
+      wrapper = renderDateRange({
+        onChange: customOnChange,
+        value: ['2012-12-12', '2012-12-13']
+      }, mount);
+      wrapper.setProps({ value: [] });
+      wrapper.update();
+      startInput = wrapper.find(BaseDateInput).at(0);
+      endInput = wrapper.find(BaseDateInput).at(1);
+      expect(startInput.props().value).toEqual('2020-01-21');
+      expect(endInput.props().value).toEqual('2020-01-21');
     });
 
     it('class names can be added to dates by passing startDateProps and endDateProps to DateRange', () => {