+++ src/components/tile/tile.stories.js
@@ -23,11 +23,11 @@ function makeStory(name, themeSelector) {
     };
 
     const tileProps = {
-      as: select('as', OptionsHelper.tileThemes, Tile.defaultProps.as, 'Tile'),
-      orientation: select('orientation', OptionsHelper.orientation, Tile.defaultProps.orientation, 'Tile'),
-      padding: select('padding', OptionsHelper.sizesTile, Tile.defaultProps.padding, 'Tile'),
-      pixelWidth: number('pixelWidth', 0, { ...percentageOpts, max: 2000 }, 'Tile'),
-      width: number('width', 0, percentageOpts, 'Tile')
+      as: select('as', OptionsHelper.tileThemes, Tile.defaultProps.as, 'Default'),
+      orientation: select('orientation', OptionsHelper.orientation, Tile.defaultProps.orientation, 'Default'),
+      padding: select('padding', OptionsHelper.sizesTile, Tile.defaultProps.padding, 'Default'),
+      pixelWidth: number('pixelWidth', 0, { ...percentageOpts, max: 2000 }, 'Default'),
+      width: number('width', 0, percentageOpts, 'Default')
     };
 
     const contentOneProps = {